email-final/requirements.txt
langgraph
langchain
langsmith
google-generativeai
google-api-python-client
google-auth
google-auth-oauthlib
python-dotenv
sqlite-utils
langchain-google-genai
